FEATURE: converts config component into shell to house form
