+++ libs/spirit-islander/feature-config/src/lib/config-form/config-form.component.html
